+++ js/main.js
@@ -1,1 +1,4 @@
-import './gallery.js';
+import { pictures } from './data.js';
+import { renderPictures } from './pictures.js';
+
+renderPictures(pictures);
