+++ js/utils.js
@@ -8,16 +8,8 @@ const getRandomArrayElement = function(arr, min = 0, max = arr.length - 1) {
   return arr[generateRandomInteger(min, max)];
 };
 
-const getPhotoByUrl = function(photos, url) {
-  for (const photo of photos) {
-    if (url.includes(photo.url)){
-      return photo;
-    }
-  }
-};
-
 const isEsc = function (evt) {
   return evt.key === 'Escape';
 };
 
-export { getRandomArrayElement, generateRandomInteger, getPhotoByUrl, isEsc };
+export { getRandomArrayElement, generateRandomInteger, isEsc };
